Fix: issue with i% and d% specifier
* changed input from long int to int

+++ main.h
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 int _printf(const char *format, ...);
 int _putchar(char c);
-char *base_converter(long int num, int base_n);
+char *base_converter(int num, int base_n);
 int arrayprinter(char *array);
 int format_printer(char j, va_list arg);
